more Optimizing!

+++ TimeTableFirstTry/TimeTableData.swift
@@ -19,9 +19,9 @@ class TimeTableData: NSManagedObject {
     @NSManaged var teacher:String
     @NSManaged var day:String
     @NSManaged var event:String
-    @NSManaged var id:Int
+    @NSManaged var id:Int64
     
-    class func createInManagedObjectContext(ManagedObjectContext moc:NSManagedObjectContext, ClassName className:String, StartTime startTime:String, EndTime endTime:String, Location location:String, Subject subject:String, Teacher teacher:String, Day day:String, Event event:String, ID id:Int) -> TimeTableData {
+    class func createInManagedObjectContext(ManagedObjectContext moc:NSManagedObjectContext, ClassName className:String, StartTime startTime:String, EndTime endTime:String, Location location:String, Subject subject:String, Teacher teacher:String, Day day:String, Event event:String, ID id:Int64) -> TimeTableData {
         let newItem = NSEntityDescription.insertNewObjectForEntityForName("TimeTableData", inManagedObjectContext: moc) as! TimeTableData
         newItem.nameOfClass = className
         newItem.startTime = startTime
